feat(projen.component.vue): migrate Vue to component.
