+++ package.json
@@ -105,7 +105,8 @@
       "packages/projen/component/pnpm-workspace",
       "packages/projen/component/unbuild",
       "packages/projen/component/vitest",
-      "packages/projen/component/linting"
+      "packages/projen/component/linting",
+      "packages/projen/component/vue"
     ]
   },
   "//": "~~ Generated by projen. To modify, edit .projenrc.ts and run \"npx projen\"."
