+++ packages/projen/component/vue/.gitattributes
@@ -1,0 +1,18 @@
+# ~~ Generated by projen. To modify, edit .projenrc.js and run "npx projen".
+
+/.eslintrc.json linguist-generated
+/.gitattributes linguist-generated
+/.gitignore linguist-generated
+/.npmignore linguist-generated
+/.prettierignore linguist-generated
+/.prettierrc.json linguist-generated
+/.projen/** linguist-generated
+/.projen/deps.json linguist-generated
+/.projen/files.json linguist-generated
+/.projen/tasks.json linguist-generated
+/build.config.ts linguist-generated
+/LICENSE linguist-generated
+/package.json linguist-generated
+/pnpm-lock.yaml linguist-generated
+/tsconfig.dev.json linguist-generated
+/tsconfig.json linguist-generated
