+++ packages/projen/component/vue/.npmignore
@@ -1,0 +1,16 @@
+# ~~ Generated by projen. To modify, edit .projenrc.js and run "npx projen".
+/.projen/
+/test/
+/src/
+!/dist/
+!/dist/**/*.js
+!/dist/**/*.d.ts
+dist
+/.github/
+/.vscode/
+/.idea/
+/.projenrc.js
+tsconfig.tsbuildinfo
+/.eslintrc.json
+/tsconfig.json
+/tsconfig.dev.json
