+++ packages/projen/component/vue/.prettierignore
@@ -1,0 +1,1 @@
+# ~~ Generated by projen. To modify, edit .projenrc.js and run "npx projen".
