+++ packages/projen/component/vue/README.md
@@ -1,0 +1,1 @@
+# replace this
