+++ packages/projen/component/vue/build.config.ts
@@ -1,0 +1,6 @@
+// ~~ Generated by projen. To modify, edit .projenrc.js and run "npx projen".
+
+import { defineBuildConfig, type BuildConfig } from 'unbuild'
+
+const config: BuildConfig = {}
+export default defineBuildConfig(config)
