+++ packages/projen/component/vue/package.json
@@ -1,0 +1,74 @@
+{
+  "name": "@arroyodev-llc/projen.component.vue",
+  "scripts": {
+    "build": "npx projen build",
+    "bump": "npx projen bump",
+    "compile": "npx projen compile",
+    "default": "npx projen default",
+    "eslint": "npx projen eslint",
+    "package": "npx projen package",
+    "post-compile": "npx projen post-compile",
+    "post-upgrade": "npx projen post-upgrade",
+    "pre-compile": "npx projen pre-compile",
+    "release": "npx projen release",
+    "stub": "npx projen stub",
+    "test": "npx projen test",
+    "unbump": "npx projen unbump",
+    "upgrade": "npx projen upgrade",
+    "watch": "npx projen watch",
+    "projen": "npx projen"
+  },
+  "author": {
+    "name": "arroyoDev-LLC",
+    "email": "support@arroyodev.com",
+    "url": "https://arroyodev.com",
+    "organization": true
+  },
+  "devDependencies": {
+    "@types/node": "^16",
+    "@typescript-eslint/eslint-plugin": "^5",
+    "@typescript-eslint/parser": "^5",
+    "eslint": "^8",
+    "eslint-config-prettier": "^8.8.0",
+    "eslint-import-resolver-node": "^0.3.7",
+    "eslint-import-resolver-typescript": "^3.5.5",
+    "eslint-plugin-import": "^2.27.5",
+    "eslint-plugin-prettier": "^4.2.1",
+    "npm-check-updates": "^16",
+    "prettier": "^2.8.8",
+    "projen": "0.71.7",
+    "standard-version": "^9",
+    "typescript": "^5",
+    "unbuild": "*"
+  },
+  "peerDependencies": {
+    "projen": "^0.71.7"
+  },
+  "dependencies": {
+    "@arroyodev-llc/projen.component.linting": "workspace:*",
+    "@arroyodev-llc/projen.component.typescript-source-file": "workspace:*",
+    "@arroyodev-llc/projen.component.unbuild": "workspace:*",
+    "@arroyodev-llc/projen.component.vitest": "workspace:*",
+    "@arroyodev-llc/utils.projen": "workspace:*",
+    "ts-morph": "^18.0.0"
+  },
+  "pnpm": {},
+  "main": "./dist/index.cjs",
+  "license": "Apache-2.0",
+  "publishConfig": {
+    "access": "public"
+  },
+  "version": "0.0.0",
+  "types": "./dist/index.d.ts",
+  "type": "module",
+  "sideEffects": false,
+  "module": "./dist/index.mjs",
+  "exports": {
+    ".": {
+      "import": "./dist/index.mjs",
+      "types": "./dist/index.d.ts",
+      "require": "./dist/index.cjs"
+    }
+  },
+  "//": "~~ Generated by projen. To modify, edit .projenrc.js and run \"npx projen\"."
+}
