+++ packages/projen/component/vue/tsconfig.dev.json
@@ -1,0 +1,17 @@
+// ~~ Generated by projen. To modify, edit .projenrc.js and run "npx projen".
+{
+  "extends": "./tsconfig.json",
+  "compilerOptions": {
+    "outDir": "dist"
+  },
+  "include": [
+    "src/*.ts",
+    "src/**/*.ts",
+    "*.ts",
+    "**/*.ts",
+    "build.config.ts"
+  ],
+  "exclude": [
+    "node_modules"
+  ]
+}
