+++ packages/projen/component/vue/tsconfig.json
@@ -1,0 +1,33 @@
+// ~~ Generated by projen. To modify, edit .projenrc.js and run "npx projen".
+{
+  "extends": [
+    "../../../../tsconfig.base.json",
+    "../../../../tsconfig.esm.json",
+    "../../../../tsconfig.bundler.json"
+  ],
+  "compilerOptions": {
+    "outDir": "dist",
+    "paths": {
+      "@arroyodev-llc/utils.projen": [
+        "../../../utils/projen/src/index"
+      ],
+      "@arroyodev-llc/projen.component.linting": [
+        "../linting/src/index"
+      ],
+      "@arroyodev-llc/projen.component.unbuild": [
+        "../unbuild/src/index"
+      ],
+      "@arroyodev-llc/projen.component.vitest": [
+        "../vitest/src/index"
+      ],
+      "@arroyodev-llc/projen.component.typescript-source-file": [
+        "../typescript-source-file/src/index"
+      ]
+    }
+  },
+  "include": [
+    "src/*.ts",
+    "src/**/*.ts"
+  ],
+  "exclude": []
+}
